+++ puppet/modules/vendors/postgresql/lib/puppet/provider/postgresql_psql/ruby.rb
@@ -16,9 +16,15 @@ Puppet::Type.type(:postgresql_psql).provide(:ruby) do
       return nil
     end
 
-    output, status = run_unless_sql_command(resource[:unless])
+    if Puppet::PUPPETVERSION.to_f < 4
+      output, status = run_unless_sql_command(resource[:unless])
+    else
+      output = run_unless_sql_command(resource[:unless])
+      status = output.exitcode
+    end
 
     if status != 0
+      puts status
       self.fail("Error evaluating 'unless' clause: '#{output}'")
     end
     result_count = output.strip.to_i
@@ -51,16 +57,36 @@ Puppet::Type.type(:postgresql_psql).provide(:ruby) do
   end
 
   def run_sql_command(sql)
+    if resource[:search_path]
+      sql = "set search_path to #{Array(resource[:search_path]).join(',')}; #{sql}"
+    end
+
     command = [resource[:psql_path]]
     command.push("-d", resource[:db]) if resource[:db]
     command.push("-t", "-c", sql)
 
     if resource[:cwd]
       Dir.chdir resource[:cwd] do
-        Puppet::Util::SUIDManager.run_and_capture(command, resource[:psql_user], resource[:psql_group])
+        run_command(command, resource[:psql_user], resource[:psql_group])
       end
     else
-      Puppet::Util::SUIDManager.run_and_capture(command, resource[:psql_user], resource[:psql_group])
+      run_command(command, resource[:psql_user], resource[:psql_group])
+    end
+  end
+
+  def run_command(command, user, group)
+    if Puppet::PUPPETVERSION.to_f < 3.4
+      Puppet::Util::SUIDManager.run_and_capture(command, user, group)
+    else
+      output = Puppet::Util::Execution.execute(command, {
+        :uid                => user,
+        :gid                => group,
+        :failonfail         => false,
+        :combine            => true,
+        :override_locale    => true,
+        :custom_environment => {}
+      })
+      [output, $CHILD_STATUS.dup]
     end
   end