+++ puppet/modules/vendors/postgresql/spec/unit/classes/params_spec.rb
@@ -3,9 +3,10 @@ require 'spec_helper'
 describe 'postgresql::params', :type => :class do
   let :facts do
     {
-      :postgres_default_version => '8.4',
       :osfamily => 'Debian',
+      :operatingsystem => 'Debian',
+      :operatingsystemrelease => '6.0',
     }
   end
-  it { should include_class("postgresql::params") }
+  it { should contain_class("postgresql::params") }
 end